Move ConstCost struct to objects.rs.
commit-id:52109be9
